+++ test/e2e/e2e_test.go
@@ -65,6 +65,7 @@ import (
 
 	// reconfigure framework
 	_ "k8s.io/kubernetes/test/e2e/framework/debug/init"
+	_ "k8s.io/kubernetes/test/e2e/framework/todo/node/init"
 )
 
 // handleFlags sets up all flags and parses the command line.
